+++ fastapi_errors_plus/base.py
@@ -121,3 +121,4 @@ class StandardErrorDTO(BaseErrorDTO):
         }
 
 
+