+++ fastapi_errors_plus/protocol.py
@@ -49,3 +49,4 @@ class ErrorDTO(Protocol):
 
 
 
+