+++ tests/conftest.py
@@ -55,3 +55,4 @@ def multiple_errors_dto():
 
 
 
+